requirements.txt
python-telegram-bot
pandas
SQLAlchemy
python-dotenv